Tests: Move Performance Plugin Tests
